Remove boxing

+++ umlet-swing/src/main/java/com/baselet/diagram/draw/swing/DrawHandlerSwing.java
@@ -118,8 +118,8 @@ public class DrawHandlerSwing extends DrawHandler {
 			int lastIdx = points.length - 1;
 			for (int i = 0; i < points.length; i++) {
 				PointDouble p = points[i];
-				Double x = inBorderHorizontal(Double.valueOf(p.getX() * getZoom() + HALF_PX), 0);
-				Double y = inBorderVertical(Double.valueOf(p.getY() * getZoom() + HALF_PX), 0);
+				Double x = inBorderHorizontal(p.getX() * getZoom() + HALF_PX, 0);
+				Double y = inBorderVertical(p.getY() * getZoom() + HALF_PX, 0);
 				if (first) {
 					path.moveTo(x, y);
 					first = false;
